Skip CI build (Travis) for release commits
These builds (especially the one which deploys the release version) are
run locally, and does not need to be run by Travis. In addition, we try
to employ several more or less, mostly less, functioning hacks to avoid
running the deploy phase for release builds, as it has already been run
when Travis picks up the commit.